requirements.txt
annotated-types==0.7.0
anyio==4.8.0
babel==2.17.0
backrefs==5.8
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
dnspython==2.7.0
docxcompose==1.4.0
docxtpl==0.19.1
EditorConfig==0.17.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fonttools==4.56.0
ghp-import==2.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsbeautifier==1.15.4
kiwisolver==1.4.8
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.8
mkdocs-material-extensions==1.3.1
mkdocs-mermaid2-plugin==1.2.1
multiprocess==0.70.17
numpy==2.2.3
packaging==24.2
paginate==0.5.7
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pymdown-extensions==10.14.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyyaml_env_tag==0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
SALib==1.5.1
scipy==1.15.2
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.0.4
websockets==15.0.1
